Added method for checking and updating xml file. Also made a few changes to the gui, such as addind JPanels for input
