+++ rdfOut.ttl
@@ -3,150 +3,146 @@
 @prefix owl:     <http://www.w3.org/2002/07/owl#> .
 @prefix rdf:     <http://www.w3.org/1999/02/22-rdf-syntax-ns#> .
 
-<2016-03-06T01:00:00>
+<2016-05-02T08:00:00>
       <http://purl.org/dc/terms/abstract>
               "Skyet" .
 
-<2016-03-04T01:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Delvis skyet" .
-
-<2016-03-09T07:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Delvis skyet" .
-
-<2016-03-02T18:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Snø" .
-
-<2016-03-02T12:00:00>
+<2016-04-28T18:00:00>
       <http://purl.org/dc/terms/abstract>
               "Skyet" .
 
-<2016-03-06T13:00:00>
+<2016-05-04T14:00:00>
       <http://purl.org/dc/terms/abstract>
-              "Skyet" .
+              "Regnbyger" .
 
-<2016-03-07T07:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Delvis skyet" .
-
-<2016-03-07T19:00:00>
+<2016-05-04T08:00:00>
       <http://purl.org/dc/terms/abstract>
               "Delvis skyet" .
 
-<2016-03-08T01:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Delvis skyet" .
-
-<2016-03-02T00:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Sludd" .
-
-<2016-03-02T06:00:00>
+<2016-05-02T20:00:00>
       <http://purl.org/dc/terms/abstract>
               "Regn" .
 
-<2016-03-04T19:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Skyet" .
-
-<2016-03-06T07:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Skyet" .
-
-<2016-03-09T01:00:00>
+<2016-05-04T02:00:00>
       <http://purl.org/dc/terms/abstract>
               "Delvis skyet" .
 
-<2016-03-01T00:00:00>
+<2016-05-02T14:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Regn" .
+
+<2016-04-28T06:00:00>
       <http://purl.org/dc/terms/abstract>
               "Skyet" .
 
-<2016-03-04T07:00:00>
+<2016-05-06T08:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Skyet" .
+
+<2016-05-01T02:00:00>
       <http://purl.org/dc/terms/abstract>
               "Delvis skyet" .
 
-<2016-03-09T13:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Skyet" .
-
-<2016-03-04T13:00:00>
+<2016-04-28T12:00:00>
       <http://purl.org/dc/terms/abstract>
               "Delvis skyet" .
 
-<2016-03-03T13:00:00>
+<2016-05-06T14:00:00>
       <http://purl.org/dc/terms/abstract>
               "Skyet" .
 
-<2016-02-29T15:00:00>
+<2016-05-04T20:00:00>
       <http://purl.org/dc/terms/abstract>
               "Skyet" .
 
-<2016-03-01T06:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Snø" .
-
-<2016-03-03T07:00:00>
+<2016-05-06T02:00:00>
       <http://purl.org/dc/terms/abstract>
               "Skyet" .
 
-<2016-03-08T19:00:00>
+<2016-05-05T02:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Delvis skyet" .
+
+<2016-04-30T00:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Delvis skyet" .
+
+<2016-05-01T08:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Delvis skyet" .
+
+<2016-05-03T14:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Regnbyger" .
+
+<2016-05-05T20:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Skyet" .
+
+<2016-04-29T06:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Delvis skyet" .
+
+<2016-05-03T20:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Regn" .
+
+<2016-05-01T14:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Skyet" .
+
+<2016-05-03T02:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Regn" .
+
+<2016-04-28T00:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Delvis skyet" .
+
+<2016-05-01T20:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Skyet" .
+
+<2016-05-03T08:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Regnbyger" .
+
+<2016-04-29T00:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Skyet" .
+
+<2016-04-27T21:00:00>
       <http://purl.org/dc/terms/abstract>
               "Klarvær" .
 
-<2016-03-03T01:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Skyet" .
-
-<2016-03-06T19:00:00>
+<2016-04-30T08:00:00>
       <http://purl.org/dc/terms/abstract>
               "Delvis skyet" .
 
-<2016-03-08T07:00:00>
+<2016-05-05T08:00:00>
       <http://purl.org/dc/terms/abstract>
               "Delvis skyet" .
 
-<2016-03-07T01:00:00>
+<2016-05-02T02:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Regn" .
+
+<2016-04-30T14:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Regnbyger" .
+
+<2016-04-30T20:00:00>
       <http://purl.org/dc/terms/abstract>
               "Delvis skyet" .
 
-<2016-03-01T12:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Snø" .
-
-<2016-03-07T13:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Skyet" .
-
-<2016-03-01T18:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Sludd" .
-
-<2016-03-03T19:00:00>
+<2016-04-29T12:00:00>
       <http://purl.org/dc/terms/abstract>
               "Delvis skyet" .
 
-<2016-03-05T19:00:00>
+<2016-04-29T18:00:00>
+      <http://purl.org/dc/terms/abstract>
+              "Regn" .
+
+<2016-05-05T14:00:00>
       <http://purl.org/dc/terms/abstract>
               "Skyet" .
-
-<2016-03-08T13:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Delvis skyet" .
-
-<2016-03-05T13:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Skyet" .
-
-<2016-02-29T18:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Skyet" .
-
-<2016-03-05T07:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Skyet" .
-
-<2016-03-05T01:00:00>
-      <http://purl.org/dc/terms/abstract>
-              "Delvis skyet" .
