+++ src/no/uib/info216/Main.java
@@ -1,8 +1,8 @@
 package no.uib.info216;
 
+import no.uib.info216.WeatherData.RDFparser;
+
 import javax.swing.*;
-import java.awt.event.MouseEvent;
-import java.awt.event.MouseListener;
 
 /**
  * Created by fox on 2/22/16.
@@ -12,17 +12,14 @@ public class Main {
 
     public static void main(String[] args) {
 
-        //Schedule a job for the event dispatch thread:
-        //creating and showing this application's GUI.
-        SwingUtilities.invokeLater(new Runnable() {
+//        SwingUtilities.invokeLater(new Runnable() {
+//            public void run() {
+//                UIManager.put("swing.boldMetal", Boolean.FALSE);
+//                newgui.createAndShowGUI();
+//            }
+//        });
 
-            public void run() {
-                //Turn off metal's use of bold fonts
-                UIManager.put("swing.boldMetal", Boolean.FALSE);
-                newgui.createAndShowGUI();
-            }
-        });
+        RDFparser weatherData = new RDFparser();
 
-        GUI gui = new GUI();
     }
 }
