+++ src/no/uib/info216/WeatherData/RDFparser.java
@@ -1,10 +1,11 @@
-package no.uib.info216;
+package no.uib.info216.WeatherData;
 
-import com.hp.hpl.jena.mem.ArrayBunch;
 import com.hp.hpl.jena.ontology.OntModel;
 import com.hp.hpl.jena.ontology.OntModelSpec;
 import com.hp.hpl.jena.rdf.model.*;
 import com.hp.hpl.jena.vocabulary.DCTerms;
+import no.uib.info216.WeatherData.JenaTest;
+
 import java.io.*;
 import java.util.*;
 
@@ -28,12 +29,12 @@ public class RDFparser {
 
             String itemdateFrom = this.dateFrom.get(i);
             String itemnameTag = this.nameTag.get(i);
-            String itemwindTag = this.windTag.get(i);
+//            String itemwindTag = this.windTag.get(i);
 
             Resource weatherData
                     = model.createResource(itemdateFrom)
-                    .addProperty(DCTerms.abstract_, itemnameTag)
-                    .addProperty(DCTerms.abstract_, itemwindTag);
+                    .addProperty(DCTerms.abstract_, itemnameTag);
+//                    .addProperty(DCTerms.abstract_, itemwindTag);
         }
